hackathon/requirements.txt
# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0

# HTTP client
httpx>=0.25.0

# Config
python-dotenv>=1.0.0

# Database
aiosqlite>=0.19.0

# Templates & PDF
jinja2>=3.1.0
playwright>=1.40.0

# Telegram bot
python-telegram-bot>=22.0

# Text-to-Speech (free, no API key)
edge-tts>=7.0.0
